requirements.txt
Pillow==7.0.0
pluggy==0.13.1
postgres==3.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
psycopg2==2.9.1
psycopg2-binary==2.9.1
psycopg2-pool==1.1
py==1.8.1
py-moneyed==0.8.0
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
pymongo==3.10.1
pyowm==2.10.0
pyparsing==2.4.6
pyrsistent==0.15.7
pyTelegramBotAPI==3.6.6
pytest==5.4.1
pytest-cov==2.8.1
pytest-django==3.9.0
python-dateutil==2.8.0
python-decouple==3.3
python-dotenv==0.10.1
python3-openid==3.1.0
pytz==2018.7
redis==3.2.0
reportlab==3.5.34
requests==2.20.1
six==1.11.0
s3transfer==0.2.0
schema==0.7.1
semantic-version==2.5.0
Send2Trash==1.5.0
sortedcontainers==1.5.9
sqlparse==0.2.4
stripe==2.45.0
swapper==1.1.2.post1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.2
texttable==0.9.1
tornado==6.0.3
traitlets==4.3.3
tzlocal==1.4
vine==1.2.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.55.0
Werkzeug==0.14.1
whitenoise==4.1.2
widgetsnbextension==3.5.1
wrapt==1.10.11
